move zip code around
